Clean up the Flask app layout
