+++ app/__init__.py
@@ -1,0 +1,48 @@
+import logging
+from flask import Flask
+from flask_bootstrap import Bootstrap
+from flask_caching import Cache
+from flask_login import LoginManager
+from flask_mail import Mail
+from flask_migrate import Migrate
+from flask_sqlalchemy import SQLAlchemy
+from config import config
+
+bootstrap = Bootstrap()
+mail = Mail()
+db = SQLAlchemy()
+migrate = Migrate()
+cache = Cache()
+
+login_manager = LoginManager()
+login_manager.session_protection = 'strong'
+login_manager.login_view = 'auth.login'
+
+
+def create_app(config_name):
+    app = Flask(__name__)
+    app.config.from_object(config[config_name])
+    config[config_name].init_app(app)
+
+    logger = app.logger
+    logger.setLevel(logging.INFO)
+    stream_handler = logging.StreamHandler()
+    logger.addHandler(stream_handler)
+
+    bootstrap.init_app(app)
+    mail.init_app(app)
+    db.init_app(app)
+    migrate.init_app(app, db)
+    login_manager.init_app(app)
+
+    if not app.debug and not app.testing and not app.config['SSL_DISABLE']:
+        from flask_sslify import SSLify
+        sslify = SSLify(app)
+
+    from .carpool import pool_bp
+    app.register_blueprint(pool_bp)
+
+    from .auth import auth_bp
+    app.register_blueprint(auth_bp)
+
+    return app
