+++ app/auth/oauth.py
@@ -19,7 +19,7 @@ class OAuthSignIn(object):
         pass
 
     def get_callback_url(self):
-        return url_for('oauth_callback',
+        return url_for('oauth.oauth_callback',
                        provider=self.provider_name,
                        _external=True,
                        _scheme='https')
