+++ app/carpool/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+pool_bp = Blueprint('carpool', __name__)
+
+from . import views
