+++ app/templates/add_rider.html
