+++ app/templates/carpool_details.html
@@ -21,10 +21,10 @@
 </p>
 
 {% if pool.get_current_user_ride_request() %}
-<a class="btn btn-default" href="{{ url_for('cancel_carpool_rider', carpool_id=pool.id) }}" role="button">Cancel your seat request</a>
+<a class="btn btn-default" href="{{ url_for('carpool.cancel_rider', carpool_id=pool.id) }}" role="button">Cancel your seat request</a>
 
 {% elif pool.current_user_is_driver %}
-<a class="btn btn-default" href="{{ url_for('cancel_carpool', carpool_id=pool.id) }}" role="button">Cancel your carpool</a>
+<a class="btn btn-default" href="{{ url_for('carpool.cancel', carpool_id=pool.id) }}" role="button">Cancel your carpool</a>
 
 <div>
 <h3>Ride Requests</h3>
@@ -33,23 +33,23 @@
     <li>{{ request.person.name }}&nbsp;<small>
         {%- if request.status == 'requested' %}
         (Waiting for your response:&nbsp;
-        <a href="{{ url_for('modify_ride_request',
+        <a href="{{ url_for('carpool.modify_ride_request',
                             action='approve',
                             carpool_id=pool.id,
                             request_id=request.id) }}">Approve</a>,&nbsp;
-        <a href="{{ url_for('modify_ride_request',
+        <a href="{{ url_for('carpool.modify_ride_request',
                             action='deny',
                             carpool_id=pool.id,
                             request_id=request.id) }}">Deny</a>)
         {%- elif request.status == 'approved' %}
         (You approved their request:&nbsp;
-        <a href="{{ url_for('modify_ride_request',
+        <a href="{{ url_for('carpool.modify_ride_request',
                             action='deny',
                             carpool_id=pool.id,
                             request_id=request.id) }}">Deny</a>)
         {%- elif request.status == 'denied' %}
         (You denied their request:&nbsp;
-        <a href="{{ url_for('modify_ride_request',
+        <a href="{{ url_for('carpool.modify_ride_request',
                             action='approve',
                             carpool_id=pool.id,
                             request_id=request.id) }}">Approve</a>)
@@ -61,7 +61,7 @@
 </div>
 
 {% else %}
-<a class="btn btn-default" href="{{ url_for('new_carpool_rider', carpool_id=pool.id) }}" role="button">Request a seat</a>
+<a class="btn btn-default" href="{{ url_for('carpool.new_ride_request', carpool_id=pool.id) }}" role="button">Request a seat</a>
 {% endif %}
 
 {% endblock %}
