--- templates/login.html
@@ -1,11 +1,0 @@
-{%- extends "_template.html" %}
-{%- import "bootstrap/wtf.html" as wtf %}
-
-{% block site %}
-<div class='page-header'>
-    <h1>Login</h1>
-</div>
-
-<p><a href="{{ url_for('oauth_authorize', provider='facebook') }}">Login with Facebook</a></p>
-<p><a href="{{ url_for('oauth_authorize', provider='google') }}">Login with Google</a></p>
-{% endblock %}
