+++ wsgi.py
@@ -1,0 +1,3 @@
+from app import create_app
+
+app = create_app('default')
